requirements.txt
annotated-types==0.7.0
black==25.1.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
filelock==3.13.1
fsspec==2024.6.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
httplib2==0.22.0
idna==3.10
isort==6.0.1
Jinja2==3.1.4
joblib==1.5.1
MarkupSafe==2.1.5
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.3
nltk==3.9.1
numpy==2.1.2
packaging==25.0
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.8
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dotenv==1.1.1
regex==2024.11.6
requests==2.32.4
rsa==4.9.1
setuptools==70.2.0
sympy==1.13.3
torch==2.7.1+cpu
torchaudio==2.7.1+cpu
torchvision==0.22.1+cpu
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
